@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,14 +14,14 @@ dependencies:
 
   azstore:
     path: ../
-  cupertino_icons: ^1.0.2
-  modal_progress_hud: ^0.1.3
-  path_provider: ^1.6.24
-  flutter_image_compress: ^0.7.0
-  image_picker: ^0.7.4
+  cupertino_icons: ^1.0.8
+  path_provider: ^2.1.4
+  modal_progress_hud_nsn: ^0.5.1
+  flutter_image_compress: ^2.3.0
+  image_picker: ^1.1.2
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
 
